fixes issue with lists of changes inside a group
This version fixes an issue where a large number of simultaneous changes were being excluded from the output because they were notarised inside nested lists and lists of lists within the data structure downloaded from Central.

The approach to getting the list of diffs was completely rewritten to account for this.